Extract privacy project sections into getPrivacyProjectEntry feature builder
Move the privacy project section assembly and entry construction out of
the page data fetcher into a server features entry builder, matching the
getScalingProjectEntry / getZkCatalogProjectEntry pattern. Pure refactor,
no user-visible change.
